Rag/requirements.txt
groq
openai
ollama
pgvector
phidata
psycopg[binary]
pypdf
sqlalchemy
streamlit
bs4
duckduckgo-search
nest_asyncio